delete algorithm file and modified CameraFragment.kt
