--- app/src/main/java/org/gradproj/heartrate/algorithm/averPixcel.kt
@@ -1,7 +1,0 @@
-package org.gradproj.heartrate.algorithm
-
-
-// 카메라에서 받은 rgb값을 계산하는 클래스를 만들어 놓으려고 했는데 안쓰일 것 같아서 비워놓음
-class averPixcel {
-
-}
