--- app/src/main/java/org/gradproj/heartrate/algorithm/mkCk.kt
@@ -1,28 +1,0 @@
-package org.gradproj.heartrate.algorithm
-// 단순 mkck와 mk값을 구해서 wck를 구하는 클래스
-class mkCk {
-
-//    init{
-//        mk(k)
-//        ck(t)
-//        mkCk(k)
-//    }
-//    fun mkCk(i: Int) : Int {
-//        if(diffR[i]!! > 15){
-//
-//          return this.ck(t)}
-//
-//        else return 0
-//
-//    }
-//    fun mk(i: Int): Int {
-//        return if(diffR[i]!! > 15) {
-//            1
-//        } else 0
-//    }
-//    fun ck(t: Int): Int {
-//
-//        return camera.flashlihgt
-//
-//    }
-}
