+++ app/src/main/java/org/gradproj/heartrate/helper/permissionListenHelper.kt
@@ -1,0 +1,14 @@
+package org.gradproj.heartrate.helper
+
+import com.gun0912.tedpermission.PermissionListener
+
+class permissionListenHelper(var granted: Boolean) : PermissionListener {
+
+    override fun onPermissionGranted() {
+
+    }
+
+    override fun onPermissionDenied(deniedPermissions: MutableList<String>?) {
+        TODO("Not yet implemented")
+    }
+}
